hr_policy_agent/requirements.txt
fastapi
uvicorn
python-dotenv
langchain-cohere
langchain-core
langchain-community
langchain-huggingface
langgraph
typing-extensions
pydantic>=2.0
model-context-protocol>=0.3.0
a2a-sdk>=0.1.0